rename : 디렉토리 구조 변경 (dto 서브 디렉토리 삭제)
프로젝트 디렉토리 구조 간소화를 위해 불필요한 디렉토리 삭제.
