+++ src/main/java/com/grablunchtogether/controller/PlanHistoryController.java
@@ -1,7 +1,7 @@
 package com.grablunchtogether.controller;
 
 import com.grablunchtogether.config.JwtTokenProvider;
-import com.grablunchtogether.dto.plan.PlanDto;
+import com.grablunchtogether.dto.PlanDto;
 import com.grablunchtogether.service.PlanHistoryService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
