+++ src/main/java/com/grablunchtogether/controller/UserController.java
@@ -2,13 +2,13 @@ package com.grablunchtogether.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.grablunchtogether.config.JwtTokenProvider;
-import com.grablunchtogether.dto.clovaOcr.ClovaOcr;
-import com.grablunchtogether.dto.geocode.GeocodeDto;
-import com.grablunchtogether.dto.image.ImageDto;
-import com.grablunchtogether.dto.naverSms.NaverSmsDto;
-import com.grablunchtogether.dto.token.TokenDto;
-import com.grablunchtogether.dto.user.OtpDto;
-import com.grablunchtogether.dto.user.UserDto;
+import com.grablunchtogether.dto.ClovaOcr;
+import com.grablunchtogether.dto.GeocodeDto;
+import com.grablunchtogether.dto.ImageDto;
+import com.grablunchtogether.dto.NaverSmsDto;
+import com.grablunchtogether.dto.TokenDto;
+import com.grablunchtogether.dto.OtpDto;
+import com.grablunchtogether.dto.UserDto;
 import com.grablunchtogether.enums.ImageDirectory;
 import com.grablunchtogether.service.*;
 import io.swagger.annotations.Api;
