+++ src/main/java/com/grablunchtogether/dto/ClovaOcr.java
@@ -1,4 +1,4 @@
-package com.grablunchtogether.dto.clovaOcr;
+package com.grablunchtogether.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModel;
