+++ src/main/java/com/grablunchtogether/dto/RefreshTokenDto.java
@@ -1,4 +1,4 @@
-package com.grablunchtogether.dto.token;
+package com.grablunchtogether.dto;
 
 import io.swagger.annotations.ApiModel;
 import lombok.AllArgsConstructor;
