+++ src/main/java/com/grablunchtogether/service/BookmarkSpotService.java
@@ -3,7 +3,7 @@ package com.grablunchtogether.service;
 import com.grablunchtogether.domain.BookmarkSpot;
 import com.grablunchtogether.domain.MustEatPlace;
 import com.grablunchtogether.domain.User;
-import com.grablunchtogether.dto.bookmarkSpot.BookmarkSpotDto;
+import com.grablunchtogether.dto.BookmarkSpotDto;
 import com.grablunchtogether.exception.CustomException;
 import com.grablunchtogether.repository.BookmarkSpotRepository;
 import com.grablunchtogether.repository.MustEatPlaceRepository;
