+++ src/main/java/com/grablunchtogether/service/PlanHistoryService.java
@@ -3,7 +3,7 @@ package com.grablunchtogether.service;
 import com.grablunchtogether.domain.Plan;
 import com.grablunchtogether.domain.PlanHistory;
 import com.grablunchtogether.domain.User;
-import com.grablunchtogether.dto.plan.PlanDto;
+import com.grablunchtogether.dto.PlanDto;
 import com.grablunchtogether.exception.CustomException;
 import com.grablunchtogether.exception.ErrorCode;
 import com.grablunchtogether.repository.PlanHistoryRepository;
