+++ src/main/java/com/grablunchtogether/service/PlanService.java
@@ -2,7 +2,7 @@ package com.grablunchtogether.service;
 
 import com.grablunchtogether.domain.Plan;
 import com.grablunchtogether.domain.User;
-import com.grablunchtogether.dto.plan.PlanDto;
+import com.grablunchtogether.dto.PlanDto;
 import com.grablunchtogether.exception.CustomException;
 import com.grablunchtogether.exception.ErrorCode;
 import com.grablunchtogether.repository.PlanRepository;
