+++ src/main/java/com/grablunchtogether/service/SMSApiService.java
@@ -3,8 +3,8 @@ package com.grablunchtogether.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.grablunchtogether.domain.User;
-import com.grablunchtogether.dto.naverSms.NaverSmsDto;
-import com.grablunchtogether.dto.plan.PlanDto;
+import com.grablunchtogether.dto.NaverSmsDto;
+import com.grablunchtogether.dto.PlanDto;
 import com.grablunchtogether.exception.CustomException;
 import com.grablunchtogether.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
