+++ src/main/java/com/grablunchtogether/service/UserReviewService.java
@@ -13,11 +13,10 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static com.grablunchtogether.dto.userReview.UserReviewDto.Dto;
-import static com.grablunchtogether.dto.userReview.UserReviewDto.UserReviewRequest;
+import static com.grablunchtogether.dto.UserReviewDto.Dto;
+import static com.grablunchtogether.dto.UserReviewDto.UserReviewRequest;
 import static com.grablunchtogether.exception.ErrorCode.*;
 
 @RequiredArgsConstructor
